# docs/openapi/tenant-openapi.yaml
openapi: 3.1.0
info:
  title: CloudAPI Tenant
  description: '

    Welcome to CloudAPI Tenant!


    For detailed guidance on using the API, please visit our official documentation:

    https://www.didx.co.za/acapy-cloud/index.html

'
  version: 5.0.0-rc5
servers:
  - url: /tenant
paths:
...
  /v1/definitions/credentials:
    post:
      tags:
        - definitions
      summary: Create a new Credential Definition
      description: "Create a credential definition\n---\nOnly issuers can create credential definitions.\n\nA credential definition essentially builds off a schema, which defines the attributes that can belong to a\ncredential, and it specifies whether credentials using this definition are revocable or not.\n\n**NB**: If revocation is requested (`\"support_revocation\": true`), then revocation registries will be created.\nThe creation of these revocation registries can take up to one minute.\n\nRequest Body:\n---\n    body: CreateCredentialDefinition\n        schema_id: str\n            The id of the schema to use for this credential definition.\n        tag: str\n            The label to use for the credential definition.\n        support_revocation: bool\n            Whether you want credentials using this definition to be revocable or not\n\nReturns\n-------\n    CredentialDefinition\n        The created credential definition"
      operationId: create_credential_definition_v1_definitions_credentials_post
      security:
        - APIKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCredentialDefinition'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialDefinition'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - definitions
      summary: Get Created Credential Definitions
      description: "Get credential definitions created by the tenant\n---\nThis endpoint returns all credential definitions created by the tenant. Only issuers can create\ncredential definitions, and so only issuers will get results from this endpoint.\n\nThe results can be filtered by the parameters listed below.\n\nParameters (Optional):\n---\n    issuer_did: str\n    credential_definition_id: str\n    schema_id: str\n    schema_issuer_did: str\n    schema_version: str\n\nReturns\n-------\n    List[CredentialDefinition]\n        A list of created credential definitions"
      operationId: get_credential_definitions_v1_definitions_credentials_get
      security:
        - APIKeyHeader: []
      parameters:
        - name: issuer_did
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Issuer Did
        - name: credential_definition_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Credential Definition Id
        - name: schema_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Schema Id
        - name: schema_issuer_did
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Schema Issuer Did
        - name: schema_name
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Schema Name
        - name: schema_version
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Schema Version
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CredentialDefinition'
                title: Response Get Credential Definitions V1 Definitions Credentials Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/definitions/credentials/{credential_definition_id}:
    get:
      tags:
        - definitions
      summary: Get a Credential Definition
      description: "Get credential definition by id\n---\nThis endpoint returns information for a credential definition.\n\nAnyone can call this, whether they created the requested credential definition or not.\nPractically it will just reveal the schema that was used for the credential definition.\n\nParameters\n----------\n    credential_definition_id: str\n        credential definition id\n\nReturns\n-------\n    CredentialDefinition\n        The credential definition"
      operationId: get_credential_definition_by_id_v1_definitions_credentials__credential_definition_id__get
      security:
        - APIKeyHeader: []
      parameters:
        - name: credential_definition_id
          in: path
          required: true
          schema:
            type: string
            title: Credential Definition Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialDefinition'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/issuer/credentials:
    post:
      tags:
        - issuer
      summary: Send Holder a Credential
      description: "Create and send a credential, automating the issuer-side flow\n---\nNB: Only a tenant with the issuer role can send credentials.\n\nWhen sending a credential, the credential type must be one of `anoncreds` or `ld_proof`.\n```json\n{\n    \"anoncreds_credential_detail\": {...},\n    \"ld_credential_detail\": {...},\n    \"save_exchange_record\": true, <-- Whether the credential exchange record should be saved on completion.\n    \"connection_id\": \"string\", <-- The issuer's reference to the connection they want to submit the credential to.\n}\n```\nSetting the `save_exchange_record` field to True will save the exchange record after the credential is accepted.\nThe default behaviour is to only save exchange records while they are in pending state.\n\nFor a detailed technical specification of the credential issuing process, refer to the [Aries Issue Credential v2\nRFC](https://github.com/hyperledger/aries-rfcs/blob/main/features/0453-issue-credential-v2/README.md).\n\nRequest Body:\n---\n    credential: SendCredential\n        The payload for sending a credential\n\nReturns\n-------\n    CredentialExchange\n        A record of this credential exchange"
      operationId: send_credential_v1_issuer_credentials_post
      security:
        - APIKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendCredential'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialExchange'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - issuer
      summary: Fetch Credential Exchange Records
      description: "Get a list of credential exchange records\n---\nBoth holders and issuers can call this endpoint, because they each have their own records of a credential exchange.\n\nThese records contain information about the credentials issued to a holder, such as the current state of the\nexchange, and other metadata such as the `connection_id` that a credential was submit to (if an issuer) or received\nfrom (if a holder). Each record in the list is related to a single credential exchange flow.\n\nNB: An issuer and a holder will have distinct credential exchange ids, despite referring to the same exchange.\nThe `thread_id` is the only record attribute that will be the same for the holder and the issuer.\n\nAn exchange record will automatically be deleted after a flow completes (i.e. when state is 'done'),\nunless the `save_exchange_record` was set to true.\n\nThe following parameters can be set to filter the fetched exchange records: connection_id, role, state, thread_id.\n\nParameters (Optional):\n---\n    limit: int - The maximum number of records to retrieve\n    offset: int - The offset to start retrieving records from\n    descending: bool - Whether to return results in descending order. Results are ordered by record created time.\n    connection_id: str\n    role: Role: \"issuer\", \"holder\"\n    state: State: \"proposal-sent\", \"proposal-received\", \"offer-sent\", \"offer-received\",\n                             \"request-sent\", \"request-received\", \"credential-issued\", \"credential-received\",\n                             \"credential-revoked\", \"abandoned\", \"done\"\n    thread_id: UUID\n\nReturns\n-------\n    List[CredentialExchange]\n        A list of credential exchange records"
      operationId: get_credentials_v1_issuer_credentials_get
      security:
        - APIKeyHeader: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
                maximum: 10000
                minimum: 1
              - type: 'null'
            description: Number of results to return
            default: 1000
            title: Limit
          description: Number of results to return
        - name: offset
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
                minimum: 0
              - type: 'null'
            description: Offset for pagination
            default: 0
            title: Offset
          description: Offset for pagination
        - name: descending
          in: query
          required: false
          schema:
            type: boolean
            description: Whether to return results in descending order. Results are ordered by record created time.
            default: true
            title: Descending
          description: Whether to return results in descending order. Results are ordered by record created time.
        - name: connection_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Connection Id
        - name: role
          in: query
          required: false
          schema:
            anyOf:
              - enum:
                  - issuer
                  - holder
                type: string
              - type: 'null'
            title: Role
        - name: state
          in: query
          required: false
          schema:
            anyOf:
              - enum:
                  - proposal-sent
                  - proposal-received
                  - offer-sent
                  - offer-received
                  - request-sent
                  - request-received
                  - credential-issued
                  - credential-received
                  - credential-revoked
                  - abandoned
                  - done
                  - deleted
                type: string
              - type: 'null'
            title: State
        - name: thread_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
                format: uuid
              - type: 'null'
            title: Thread Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CredentialExchange'
                title: Response Get Credentials V1 Issuer Credentials Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/issuer/credentials/{credential_exchange_id}:
    get:
      tags:
        - issuer
      summary: Fetch a single Credential Exchange Record
      description: "Get a credential exchange record by credential id\n---\nBoth holders and issuers can call this endpoint, because they each have their own records of a credential exchange.\n\nThese records contain information about the credentials issued to a holder, such as the current state of the\nexchange, and other metadata such as the `connection_id` that a credential was submit to (if an issuer) or received\nfrom (if a holder).\n\nNB: An issuer and a holder will have distinct credential exchange ids, despite referring to the same exchange.\nThe `thread_id` is the only record attribute that will be the same for the holder and the issuer.\n\nAn exchange record will, by default, automatically be deleted after a flow completes (i.e. when state is 'done'),\nunless the `save_exchange_record` was set to true, or the wallet is configured to preserve records by default.\n\nThe following parameters can be set to filter the fetched exchange records: connection_id, role, state, thread_id.\n\nParameters\n----------\n    credential_exchange_id: str\n        The identifier of the credential exchange record that you want to fetch\n\nReturns\n-------\n    CredentialExchange\n        The credential exchange record"
      operationId: get_credential_v1_issuer_credentials__credential_exchange_id__get
      security:
        - APIKeyHeader: []
      parameters:
        - name: credential_exchange_id
          in: path
          required: true
          schema:
            type: string
            title: Credential Exchange Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialExchange'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - issuer
      summary: Delete an Exchange Record
      description: "Delete a credential exchange record\n---\nThis will remove a specific credential exchange from your storage records.\n\nParameters\n----------\n    credential_exchange_id: str\n        The identifier of the credential exchange record that you want to delete\n\nReturns\n-------\n    status_code: 204"
      operationId: remove_credential_exchange_record_v1_issuer_credentials__credential_exchange_id__delete
      security:
        - APIKeyHeader: []
      parameters:
        - name: credential_exchange_id
          in: path
          required: true
          schema:
            type: string
            title: Credential Exchange Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/verifier/proofs/{proof_id}/credentials:
    get:
      tags:
        - verifier
      summary: Get Matching Credentials for a Proof
      description: "Get matching credentials for a presentation exchange\n---\nThis endpoint returns a list of possible credentials that the prover can use to respond to a given proof request.\n\nThe `presentation_referents` field (in the response) indicates which of the fields\nin the proof request that credential satisfies.\n\nParameters\n----------\n    proof_id: str\n        The relevant proof exchange ID for the prover\n    referent: Optional str\n        The presentation_referent of the proof to match, comma separated str of presentation_referents\n    limit: Optional int\n        The number of credentials to fetch\n    offset: Optional int\n        The index to start fetching credentials from\n\nReturns\n-------\n    List[CredPrecis]\n        A list of applicable credentials"
      operationId: get_credentials_by_proof_id_v1_verifier_proofs__proof_id__credentials_get
      security:
        - APIKeyHeader: []
      parameters:
        - name: proof_id
          in: path
          required: true
          schema:
            type: string
            title: Proof Id
        - name: referent
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Referent
        - name: limit
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
                maximum: 10000
                minimum: 1
              - type: 'null'
            description: Number of results to return
            default: 1000
            title: Limit
          description: Number of results to return
        - name: offset
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
                minimum: 0
              - type: 'null'
            description: Offset for pagination
            default: 0
            title: Offset
          description: Offset for pagination
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CredPrecis'
                title: Response Get Credentials By Proof Id V1 Verifier Proofs  Proof Id  Credentials Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/wallet/credentials:
    get:
      tags:
        - wallet
      summary: Fetch a list of credentials from the wallet
      description: "Fetch a list of credentials from the wallet\n---\n\nThe `wql` (Wallet Query Language) parameter can be used to filter credentials returned from the wallet.\n\nThe following string will look for the credential with the attribute `age` with value `21`:\n\n    {\"attr::age::value\": \"21\"}\n\nOptional Parameters:\n---\n    limit: int\n        The number of records to return.\n    offset: int\n        The number of records to skip before starting to return records.\n    wql: str\n        A WQL query to filter records.\n\nReturns\n-------\n    CredInfoList\n        A list of credential records."
      operationId: list_credentials_v1_wallet_credentials_get
      security:
        - APIKeyHeader: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
                maximum: 10000
                minimum: 1
              - type: 'null'
            description: Number of results to return
            default: 1000
            title: Limit
          description: Number of results to return
        - name: offset
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
                minimum: 0
              - type: 'null'
            description: Offset for pagination
            default: 0
            title: Offset
          description: Offset for pagination
        - name: wql
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Wql
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredInfoList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/wallet/credentials/{credential_id}:
    get:
      tags:
        - wallet
      summary: Fetch a credential by ID
      description: "Fetch a specific credential by ID\n---\n\nParameters\n----------\n    credential_id: str\n        The ID of the credential to fetch.\n\nReturns\n-------\n    CredInfo\n        The credential record."
      operationId: get_credential_record_v1_wallet_credentials__credential_id__get
      security:
        - APIKeyHeader: []
      parameters:
        - name: credential_id
          in: path
          required: true
          schema:
            type: string
            title: Credential Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredInfo'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - wallet
      summary: Delete a credential by ID
      description: "Remove a specific credential from the wallet by ID\n---\n\nParameters\n----------\n    credential_id: str\n        The ID of the credential to delete.\n\nReturns\n-------\n    status_code: 204"
      operationId: delete_credential_v1_wallet_credentials__credential_id__delete
      security:
        - APIKeyHeader: []
      parameters:
        - name: credential_id
          in: path
          required: true
          schema:
            type: string
            title: Credential Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/wallet/credentials/{credential_id}/mime-types:
    get:
      tags:
        - wallet
      summary: Retrieve attribute MIME types of a credential
      description: "Retrieve attribute MIME types of a specific credential by ID\n---\n\nParameters\n----------\n    credential_id: str\n        The ID of the credential to fetch attribute MIME types for.\n\nReturns\n-------\n    AttributeMimeTypesResult\n        The attribute MIME types of the credential."
      operationId: get_credential_mime_types_v1_wallet_credentials__credential_id__mime_types_get
      security:
        - APIKeyHeader: []
      parameters:
        - name: credential_id
          in: path
          required: true
          schema:
            type: string
            title: Credential Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttributeMimeTypesResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AttributeMimeTypesResult:
      properties:
        results:
          anyOf:
            - additionalProperties:
                type: string
              type: object
            - type: 'null'
          title: Results
      type: object
      title: AttributeMimeTypesResult
      description: AttributeMimeTypesResult
...
    CreateCredentialDefinition:
      properties:
        schema_id:
          type: string
          title: Schema Id
          examples:
            - CXQseFxV34pcb8vf32XhEa:2:test_schema:0.3
        tag:
          type: string
          title: Tag
          examples:
            - default
        support_revocation:
          type: boolean
          title: Support Revocation
          default: false
      type: object
      required:
        - schema_id
        - tag
      title: CreateCredentialDefinition
...
    CredInfoList:
      properties:
        results:
          anyOf:
            - items:
                $ref: '#/components/schemas/CredInfo'
              type: array
            - type: 'null'
          title: Results
      type: object
      title: CredInfoList
...
    CredentialDefinition:
      properties:
        id:
          type: string
          title: Id
          examples:
            - 5Q1Zz9foMeAA8Q7mrmzCfZ:3:CL:7:default
        tag:
          type: string
          title: Tag
          examples:
            - default
        schema_id:
          type: string
          title: Schema Id
          examples:
            - CXQseFxV34pcb8vf32XhEa:2:test_schema:0.3
      type: object
      required:
        - id
        - tag
        - schema_id
      title: CredentialDefinition
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: x-api-key